+++ packages/terra-search-field/CHANGELOG.md
@@ -6,6 +6,14 @@ Unreleased
 ### Fixed
 * Fixed search input rendering on Mobile Safari.
 
+### Changed
+* Update nightwatch tests and test scripts.
+
+1.6.0 - (August 31, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.5.0 - (August 15, 2017)
 ------------------
 ### Changed